Refactored handlers: structs in packages
Moved from single functions to different structs in different packages,
so they can contain context and can be loaded in a cleaner way.

The mailqueue is currently disabled as it must be implemented into the HandlerMechanism in the future
